XMLHttpRequest/send-non-same-origin.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: send() - non same-origin</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <base href="http://test2.w3.org/">
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#cross-origin-request-steps" data-tested-assertations="/following::DL[2]/DT[1] /following::DL[2]/DD[1]" />  
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#cross-origin-request-event-rules" data-tested-assertations="/following::DL[1]/DT[2] /following::DL[1]/DD[2]" />  
  </head>
  <body>
    <div id="log"></div>
    <script>
      function url(url) {
        test(function() {
          var client = new XMLHttpRequest()
          client.open("GET", url, false)
          assert_throws("NetworkError", function() { client.send() })
        }, document.title + " (" + url + ")")
      }
      url("mailto:test@example.org")
      url("tel:+31600000000")
      url("about:blank")
      url("data:text/html,<!doctype html><p>?")
      url("http://test2.w3.org/")
      url("javascript:alert('FAIL')")
      url("folder.txt")
    </script>
  </body>
</html>
