LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/send-non-same-origin-expected.txt

CONSOLE MESSAGE: line 21: XMLHttpRequest cannot load mailto:test@example.org. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 21: XMLHttpRequest cannot load tel:+31600000000. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 21: XMLHttpRequest cannot load http://127.0.0.1:8800/. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: line 21: XMLHttpRequest cannot load javascript:alert('FAIL'). Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 21: XMLHttpRequest cannot load http://127.0.0.1:8800/folder.txt. Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: line 21: XMLHttpRequest cannot load about:blank. Cross origin requests are only supported for HTTP.
CONSOLE MESSAGE: line 21: XMLHttpRequest cannot load blob:bogusidentifier. Cross origin requests are only supported for HTTP.

PASS XMLHttpRequest: send() - non same-origin (mailto:test@example.org) 
PASS XMLHttpRequest: send() - non same-origin (tel:+31600000000) 
PASS XMLHttpRequest: send() - non same-origin (http://127.0.0.1:8800) 
PASS XMLHttpRequest: send() - non same-origin (javascript:alert('FAIL')) 
PASS XMLHttpRequest: send() - non same-origin (folder.txt) 
PASS XMLHttpRequest: send() - non same-origin (about:blank) 
PASS XMLHttpRequest: send() - non same-origin (blob:bogusidentifier) 
